Functionally interactive map
Clicking on state/county adds or removes that item on the filter multiselects.

Changing polygon colors on the fly is disabled. It was working on polygon click, but not on filter selection change. So while the results were correct, the colors could get out of sync if a user was using both map clicks and location/county dropdown selections.

There were additional complications when a county would still be highlighted red, but wouldn't be in the county filter any more because its state isn't in the filter.